+++ src/button/button.hpp
@@ -1,8 +1,8 @@
 #pragma once
 #include "settings.h"
-#if BUTTONS
+#if BUTTON_SUM
 #include <Arduino.h>
-PROGMEM_INT buttonPins[BUTTONS] = {BTNS};
+PROGMEM_INT buttonPins[BUTTON_SUM] = {BTNS};
 
 class button
 {